refactor(ari): extract error to separate module
